tools/sprite_extractor/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
